dev: add dependency from selected project to /project
